+++ src/Oriflame.PolicyBuilder.Xml/Builders/Sections/BackendSectionPolicyBuilder.SetQueryParameter.cs
@@ -7,6 +7,12 @@ namespace Oriflame.PolicyBuilder.Xml.Builders.Sections
     public partial class BackendSectionPolicyBuilder : SectionBuilderBase<IBackendSectionPolicyBuilder>, IBackendSectionPolicyBuilder
     {
         /// <inheritdoc />
+        public IBackendSectionPolicyBuilder DeleteQueryParameter(string name)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        /// <inheritdoc />
         public virtual IBackendSectionPolicyBuilder SetQueryParameter(string name, string value, ExistsAction? existsAction)
         {
             return AddPolicyDefinition(new SetQueryParameter(name, value, existsAction));