+++ src/Oriflame.PolicyBuilder.Xml/Definitions/Common/DeleteQueryParameter.cs
@@ -1,0 +1,18 @@
+﻿using System.Xml.Linq;
+
+namespace Oriflame.PolicyBuilder.Xml.Definitions.Common
+{
+    public class DeleteQueryParameter : PolicyXmlBase
+    {
+        public DeleteQueryParameter(string name) : base("set-query-parameter")
+        {
+            Attributes.Add("name", name);
+            Attributes.Add("exists-action", "delete");
+        }
+
+        public override XNode GetXml()
+        {
+            return CreateElement();
+        }
+    }
+}