+++ src/Oriflame.PolicyBuilder.Xml/Definitions/Common/ForwardRequest.cs
@@ -1,20 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
 using Oriflame.PolicyBuilder.Policies.Extensions;
 
 namespace Oriflame.PolicyBuilder.Xml.Definitions.Common
 {
     public class ForwardRequest : PolicyXmlBase
     {
-        public ForwardRequest(TimeSpan? timeout) : this(timeout?.GetSeconds())
+        public ForwardRequest(IDictionary<string, string> attributes) : base("forward-request", attributes)
         {
-        }
-
-        public ForwardRequest(string timeoutValue) : base("forward-request")
-        {
-            if (!string.IsNullOrEmpty(timeoutValue))
-            {
-                Attributes.Add("timeout", timeoutValue);
-            }
         }
     }
 }