+++ src/Oriflame.PolicyBuilder.Xml/Definitions/Common/SetQueryParameter.cs
@@ -11,14 +11,25 @@ namespace Oriflame.PolicyBuilder.Xml.Definitions.Common
 
         public SetQueryParameter(string name, string value, ExistsAction? existsAction) : base("set-query-parameter")
         {
-            _value = new Value(value);
+#pragma warning disable CS0618 // Type or member is obsolete
+            if (existsAction == ExistsAction.Delete && value != null)
+            {
+                throw new System.ArgumentException("Parameter marked for deletion cannot have value(s) specified.");
+            }
+#pragma warning restore CS0618 // Type or member is obsolete
+
+            if (value != null)
+            {
+                _value = new Value(value);
+            }
+
             Attributes.Add("name", name);
             Attributes.Add("exists-action", ExistsActionMapper.Map(existsAction));
         }
 
         public override XNode GetXml()
         {
-            return CreateElement( _value.GetXml());
+            return _value == null ? CreateElement() : CreateElement(_value.GetXml());
         }
     }
 }