+++ tests/Oriflame.PolicyBuilder.Xml.Tests/Definitions/Common/ForwardRequestsTests.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using System.Collections.Generic;
 using FluentAssertions;
 using Oriflame.PolicyBuilder.Xml.Definitions.Common;
 using Xunit;
@@ -7,13 +7,12 @@ namespace Oriflame.PolicyBuilder.Xml.Tests.Definitions.Common
 {
     public class ForwardRequestsTests
     {
-        [Theory]
-        [InlineData(5)]
-        public void CreatesCorrectPolicy(int timeoutSeconds)
+        [Fact]
+        public void CreatesCorrectPolicy()
         {
-            var basePolicy = new ForwardRequest(TimeSpan.FromSeconds(timeoutSeconds));
+            var basePolicy = new ForwardRequest(new Dictionary<string, string>());
             var xml = basePolicy.GetXml().ToString();
-            xml.Should().Be($"<forward-request timeout=\"{timeoutSeconds}\" />");
+            xml.Should().Be($"<forward-request />");
         }
     }
 }